requirements.txt
openai==1.51.2
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
fastapi==0.115.2
uvicorn==0.32.0
